fix timer
